+++ src/features/address/api/getAddresses.ts
@@ -1,0 +1,6 @@
+import { API_END_POINT, httpClient } from "@/shared";
+
+export const getAddresses = async () => {
+  const response = await httpClient.get(API_END_POINT.ADDRESS);
+  return response.data;
+};
