+++ src/features/address/api/getCustomerKey.ts
@@ -1,0 +1,6 @@
+import { API_END_POINT, httpClient } from "@/shared";
+
+export const getCustomerKey = async () => {
+  const response = await httpClient.get(API_END_POINT.CUSTOMER_KEY);
+  return response.data;
+};
