+++ src/features/address/model/index.ts
@@ -1,0 +1,4 @@
+export { useDeleteAddress } from "./useDeleteAddress";
+export { useEditAddress } from "./useEditAddress";
+export { useGetAddresses } from "./useGetAddresses";
+export { usePostAddress } from "./usePostAddress";
