+++ src/pages/payment/ui/PaymentAddressList.tsx
@@ -1,7 +1,6 @@
 import { useEffect, useState } from "react";
 import { useLocation, useNavigate, useParams } from "react-router-dom";
 import { Layout } from "@/app/layout/index";
-import { useGetAddresses } from "@/components/address/queries";
 import { ADDRESS_SCRIPT_URL } from "@/constants/address";
 import { Button } from "@/shared";
 import rocation_off from '@/shared/assets/icons/rocation_off.svg';
@@ -10,6 +9,7 @@ import { ROUTES } from "@/shared/constants/routes";
 import { FaCheck } from "react-icons/fa6";
 import { IoIosSearch } from "react-icons/io";
 import type { IAddressWithId } from "@/@types/Address";
+import { useGetAddresses } from "@/features/address/model";
 
 export const PaymentAddressList = () => {
   const navigate = useNavigate();
